refactor Partner DAOs
